Call onCompleted with data from cache (#2190)
* Call onCompleted with data from cache

* Trigger `onCompleted` / `onError` callbacks after render

Adjusts `<Query />` `onCompleted` and `onError` callbacks to be
triggered via the `componentDidUpdate` lifecycle method. This
ensures these callbacks can be used when data is fetched over
the network, and when data is fetched from the local store
(previsouly these callbacks were only being triggered when data
was fetched over the network).