Replace getDataFromTree Trie class with two-layered Map.
This Trie class was only ever used to look up sequences of two elements,
where the first element was always a DocumentNode and the second element
was always a string.

PR #2753 by @mikebm highlighted the fact that the Trie was not as easy to
reuse as it was originally intended to be. By comparison, a simple
Map<DocumentNode, Map<string, QueryInfo>> data structure has better type
safety and a clearer purpose, while enabling easier extensions (just add
new fields to QueryInfo and update the makeDefaultQueryInfo function).