Make sure  `npm run watch` builds ESM/UMB bundles (#2765)
When testing new `react-apollo` changes out (during core
development), it can be useful to `npm link` `react-apollo`
into a test app. After the link is in place, the `npm run watch`
script can be used to compile typescript file changes, in watch
mode. So while working on the `react-apollo` codebase, .ts file
changes will be picked up automatically, compiled, and because of
the `npm link`, the test app will recognize that there have
been changes made to `react-apollo` and re-run with the updated
version.

Unfortunately, `npm run watch` doesn’t currently work as it’s
missing the additional post compilation steps that
`npm run compile` leverages via `postcompile`. This mean that
`npm run watch` isn’t calling `rollup` and generating the
associated ESM/UMD bundles. So when typescript changes are picked
up and compiled, the application at the other end of the link
doesn’t always recognize that `react-apollo` has been updated,
since it doesn’t see changes in the ESM/UMD bundles (if it’s
using them - e.g. `create-react-app` is).

This commit first introduces the `tsc-watch` helper utility to
make it easier to run scripts after watched typescript files are
compiled. It then wires in changes to call `npm run postcompile`
after watched typescript compilation has completed. This helps
ensure that apps at the other end of the link see changes,
since the ESM/UMD bundles are regenerated.

One thing to note is that running rollup with the current config
takes about 3 seconds. This adds a delay between making changes in
`react-apollo` and seeing them show up in the linked application.
We can reduce this time by adjusting the rollup config to
accommodate watching and only generate the bundles we need, but
given that this functionality is currently broken as is, this
commit will at least get things back to a working state.