Initial data layer abstraction
This is the beginning of a large chunk of refactoring,
intended to help abstract the core data layer functionality
out of the current React Apollo render prop based components.
The code behind the components has become a bit unwieldy
over the years, making it very easy for new bugs to be
introduced, and difficult to track down existing bugs. The
Apollo `Query`, `Mutation` and `Subscription` components
themselves also share a fair amount of similarities between
their implementation, that can be pulled out and re-used,
to help reduce bundle sizes.

This commit kick things off by extracting all non-React
dependent functionality out of the core Apollo components,
storing that functionality in new core classes. The Apollo
components themselves are left as very simple React
functional components, that leverage  Hooks, and call into
the new core classes as needed. Again, this is just the
start - more commits are coming that:

- Extract all common code out of these new core classes,
  making sure that common code is re-used appropriately
- Introduce a common API across all core classes, so their
  functionality is easier to reason about / understand
- Move core data layer handling out of the
  `@apollo/react-components` package, into `@apollo/react-hooks`
  (since the hooks package will eventually be the source
  package that drives `@apollo/react-hoc` and
  `@apollo/react-components`