Overhaul testing infra with `act` / `react-testing-library`
React Apollo is currently using 4 different React testing
libraries:

- `enzyme`
- `react-dom/test-utils`
- `react-test-renderer`
- `react-testing-library`

Needless to say, there is a substantial amount of
functionality overlap between these packages, and we don't
need to use all of them. They have been introduced over the
past couple of years as different developers have worked on
`react-apollo`, and have looked to address certain needs
that other testing packages couldn't accommodate. Luckily
with modern day React, this is no longer necessary.

This commit updates all React Apollo tests to use
`react-testing-library`, removing the use of all other React
testing approaches. We've decided to standardize on
`react-testing-library`, for the following reasons:

1. We agree with `react-testing-library`'s guiding
   principles (https://testing-library.com/docs/guiding-principles)

2. We're on the cusp of committing our Hooks work, which
   requires the use of React's `act` testing utility.
   `react-testing-library` has done an excellent job integrating
   `act` behind the scenes, so we don't have to worry about it
   (as much).

This is a large commit, as most of our tests have had to change.
This commit also includes functionality changes that were necessary
to better align ourselves with Hooks, as an outcome of `act` testing.