Wire `returnPartialData` prop up
This lines up with the Apollo Client 2.6 changes made in
https://github.com/apollographql/apollo-client/pull/4586/commits/6bbe86c413063b7ffa611abcbb24dbc07e38aa13. When performing a query that includes either more or less data
then a previous similar query cached data set, if
`returnPartialData` is set to `true`, the partially matched
cache data will be returned to the component for use. If
`returnPartialData` is `false` (the default behavior), cache
mismatching will lead to the current day behavior of returning
nothing. In a future major release, `returnPartialData` will be
enabled by default (yay!).