Initialize data as {} when using useQuery with SSR (#3175)
* Initialize data as {} when using useQuery with SSR

This is to line up with the `useQuery` initial client side `data`
value of `{}`. This changed when we moved from using the
deprecated `ObservableQuery.currentResult()` method to the newer
`ObservableQuery.getCurrentResult()` method (which doesn't
automatically convert the initial `undefined` `data` value to
`{}`).

Fixes #3143.

* Fix package lock craziness

* Move `useQuery` SSR tests to the `ssr` package

Testing SSR functionality in the `hooks` package introduces
a dependency cycle:

@apollo/react-hooks -> @apollo/react-ssr -> @apollo/react-hooks

To avoid this, let's move the `useQuery` SSR tests to the `ssr`
package.