Add `useQuery` lazy mode to defer query execution
Initial work to support deferring `useQuery` execution until
some point in the future (after `useQuery` has been initialized).

If a `lazy` option is set to a boolean, `useQuery` will return a
tuple instead of a result object. The first position of the tuple
is the standard query result object, and the second position is
a function that can be triggered to kickstart the normal
`useQuery` query cycle.

If `lazy` is `true`, the initial query result with have a
`loading` status of `false`, and a `data` value of `undefined`.
This will only change after query execution has been triggered
using the delayed start function.

If `lazy` is `false`, a tuple is still returned, but
`useQuery` will still fire immediately (similar to calling
`useQuery` without a `lazy` option).

Fixes https://github.com/apollographql/apollo-feature-requests/issues/119.